+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Change `SymbolEntryIndex.findSymbolsByClassName(...)` to return symbols.
 * Add `SymbolEntry.toString()` method that emits serialized form of symbol entry.
 * Rename `SymbolEntryIndex.findMembersByPatterns(..)` to `SymbolEntryIndex.findSymbolsByPatterns(..)`.
+* Add `SymbolEntryIndex.findSymbolsByPatterns(..)` method that accepts string patterns.
 
 ### [v0.01](https://github.com/realityforge/gir/tree/v0.01) (2018-03-14)
 [Full Changelog](https://github.com/realityforge/gir/compare/2edd3c5155779ee1a5b830a7056d374efc33002e...v0.01)